[WIP] Explore remote N5 datasets
